+++ core/src/main/java/games/cubi/raycastedantiesp/core/locatables/NettyEntityLocatable.java
@@ -1,4 +1,4 @@
-package games.cubi.locatables.minecraft;
+package games.cubi.raycastedantiesp.core.locatables;
 
 import games.cubi.locatables.MutableLocatable;
 import games.cubi.locatables.implementations.MutableLocatableImpl;
